Moved game models out to individual files
