+++ Sudoku.Web.Server/Models/GameStatisticsModel.cs
@@ -1,0 +1,11 @@
+namespace Sudoku.Web.Server.Models;
+
+/// <summary>
+/// Represents game statistics
+/// </summary>
+public record GameStatisticsModel(
+    int TotalMoves,
+    int ValidMoves,
+    int InvalidMoves,
+    TimeSpan PlayDuration,
+    double AccuracyPercentage);
