+++ Sudoku.Web.Server/Models/Requests/CellRequest.cs
@@ -1,0 +1,6 @@
+namespace Sudoku.Web.Server.Models.Requests;
+
+/// <summary>
+/// Request model for cell operations
+/// </summary>
+public record CellRequest(int Row, int Column);
