+++ Sudoku.Web.Server/Models/Requests/CreatePlayerRequest.cs
@@ -1,0 +1,6 @@
+namespace Sudoku.Web.Server.Models;
+
+/// <summary>
+/// Request model for creating a player
+/// </summary>
+public record CreatePlayerRequest(string? Alias);
