+++ Sudoku.Web.Server/Models/Requests/MoveRequest.cs
@@ -1,0 +1,6 @@
+namespace Sudoku.Web.Server.Models;
+
+/// <summary>
+/// Request model for making a move
+/// </summary>
+public record MoveRequest(int Row, int Column, int? Value);
