+++ Sudoku.Web.Server/Models/Requests/PossibleValueRequest.cs
@@ -1,0 +1,6 @@
+namespace Sudoku.Web.Server.Models.Requests;
+
+/// <summary>
+/// Request model for possible value operations
+/// </summary>
+public record PossibleValueRequest(int Row, int Column, int Value);
