+++ Sudoku.Web.Server/Models/ValidationResultModel.cs
@@ -1,0 +1,9 @@
+namespace Sudoku.Web.Server.Models;
+
+/// <summary>
+/// Represents the result of game validation
+/// </summary>
+public record ValidationResultModel(
+    bool IsValid,
+    bool IsComplete,
+    List<string> Errors);
